change config setup
